012_delivery_html/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
bandit==1.7.9
black==24.4.2
certifi==2024.6.2
click==8.1.7
coverage==7.5.3
dnspython==2.6.1
email_validator==2.1.2
fastapi==0.111.0
fastapi-cli==0.0.4
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mypy-extensions==1.0.0
numpy==2.0.0
orjson==3.10.5
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
pbr==6.0.0
platformdirs==4.2.2
pluggy==1.5.0
psycopg2==2.9.9
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
pytest==8.2.2
pytest-testdox==3.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
stevedore==5.2.0
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
